Refactor input identifier of map

+++ src/input/mapping.h
@@ -17,6 +17,7 @@
  * along with Moonlight; if not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <stdint.h>
 #include <stdbool.h>
 
 struct mapping {
@@ -31,12 +32,12 @@ struct mapping {
   short abs_dpad_x, abs_dpad_y;
   bool reverse_dpad_x, reverse_dpad_y;
 
-  short btn_south, btn_east, btn_north, btn_west;
-  short btn_select, btn_start, btn_mode;
-  short btn_thumbl, btn_thumbr;
-  short btn_tl, btn_tr, btn_tl2, btn_tr2;
+  uint32_t btn_south, btn_east, btn_north, btn_west;
+  uint32_t btn_select, btn_start, btn_mode;
+  uint32_t btn_thumbl, btn_thumbr;
+  uint32_t btn_tl, btn_tr, btn_tl2, btn_tr2;
 
-  short btn_dpad_up, btn_dpad_down, btn_dpad_left, btn_dpad_right;
+  uint32_t btn_dpad_up, btn_dpad_down, btn_dpad_left, btn_dpad_right;
 };
 
 void mapping_load(char* fileName, struct mapping* map);
